added sales_analyst_helper module and rspec hardenss complete. end iteration 2
