+++ lib/modules/repository.rb
@@ -59,7 +59,7 @@ module Repository
     single_data = find_by_id(id_to_find)
     single_data.change_attributes(hash) if single_data != nil
   end
-####
+  
   def find_all_by_price_greater_or_lesser(money, key) # key is a boolean
     @repository.find_all do |item|
       if !key
