--- lib/modules/sales_helper.rb
@@ -1,1 +1,0 @@
-
